Extract shared AI prompt helpers into utils/prompt
Move sanitizePromptInput, getTodayKey, and the ConversationTurn interface
out of trades.ts into a new server/src/utils/prompt.ts, and import them in
trades.ts and tradeHistory.ts (dropping tradeHistory's duplicate
getTodayKey). Pure refactor to share these with the upcoming
draft-rankings ask route.
